third_party/blink/web_tests/storage/websql/open-database-set-empty-version-expected.txt

CONSOLE ERROR: unable to open database, version mismatch, 'test' does not match the currentVersion of ''
This tests that calling openDatabase with an empty version string sets the current version of that database to the empty string and subsequent attempts to open the database with a different expected version throw an exception.
SUCCESS, an exception was thrown. InvalidStateError: Failed to execute 'openDatabase' on 'Window': unable to open database, version mismatch, 'test' does not match the currentVersion of ''
